Replace some code to separate files
