+++ lib/fileutils.rb
@@ -7,6 +7,8 @@ rescue LoadError
 end
 
 require_relative "fileutils/version"
+require_relative 'fileutils/stream_utils'
+require_relative 'fileutils/entry'
 
 #
 # = fileutils.rb
@@ -1156,422 +1158,8 @@ module FileUtils
 
   private
 
-  module StreamUtils_
-    private
-
-    case (defined?(::RbConfig) ? ::RbConfig::CONFIG['host_os'] : ::RUBY_PLATFORM)
-    when /mswin|mingw/
-      def fu_windows?; true end
-    else
-      def fu_windows?; false end
-    end
-
-    def fu_copy_stream0(src, dest, blksize = nil)   #:nodoc:
-      IO.copy_stream(src, dest)
-    end
-
-    def fu_stream_blksize(*streams)
-      streams.each do |s|
-        next unless s.respond_to?(:stat)
-        size = fu_blksize(s.stat)
-        return size if size
-      end
-      fu_default_blksize()
-    end
-
-    def fu_blksize(st)
-      s = st.blksize
-      return nil unless s
-      return nil if s == 0
-      s
-    end
-
-    def fu_default_blksize
-      1024
-    end
-  end
-
   include StreamUtils_
   extend StreamUtils_
-
-  class Entry_   #:nodoc: internal use only
-    include StreamUtils_
-
-    def initialize(a, b = nil, deref = false)
-      @prefix = @rel = @path = nil
-      if b
-        @prefix = a
-        @rel = b
-      else
-        @path = a
-      end
-      @deref = deref
-      @stat = nil
-      @lstat = nil
-    end
-
-    def inspect
-      "\#<#{self.class} #{path()}>"
-    end
-
-    def path
-      if @path
-        File.path(@path)
-      else
-        join(@prefix, @rel)
-      end
-    end
-
-    def prefix
-      @prefix || @path
-    end
-
-    def rel
-      @rel
-    end
-
-    def dereference?
-      @deref
-    end
-
-    def exist?
-      begin
-        lstat
-        true
-      rescue Errno::ENOENT
-        false
-      end
-    end
-
-    def file?
-      s = lstat!
-      s and s.file?
-    end
-
-    def directory?
-      s = lstat!
-      s and s.directory?
-    end
-
-    def symlink?
-      s = lstat!
-      s and s.symlink?
-    end
-
-    def chardev?
-      s = lstat!
-      s and s.chardev?
-    end
-
-    def blockdev?
-      s = lstat!
-      s and s.blockdev?
-    end
-
-    def socket?
-      s = lstat!
-      s and s.socket?
-    end
-
-    def pipe?
-      s = lstat!
-      s and s.pipe?
-    end
-
-    S_IF_DOOR = 0xD000
-
-    def door?
-      s = lstat!
-      s and (s.mode & 0xF000 == S_IF_DOOR)
-    end
-
-    def entries
-      opts = {}
-      opts[:encoding] = ::Encoding::UTF_8 if fu_windows?
-
-      files = if Dir.respond_to?(:children)
-        Dir.children(path, opts)
-      else
-        Dir.entries(path(), opts)
-           .reject {|n| n == '.' or n == '..' }
-      end
-
-      files.map {|n| Entry_.new(prefix(), join(rel(), n.untaint)) }
-    end
-
-    def stat
-      return @stat if @stat
-      if lstat() and lstat().symlink?
-        @stat = File.stat(path())
-      else
-        @stat = lstat()
-      end
-      @stat
-    end
-
-    def stat!
-      return @stat if @stat
-      if lstat! and lstat!.symlink?
-        @stat = File.stat(path())
-      else
-        @stat = lstat!
-      end
-      @stat
-    rescue SystemCallError
-      nil
-    end
-
-    def lstat
-      if dereference?
-        @lstat ||= File.stat(path())
-      else
-        @lstat ||= File.lstat(path())
-      end
-    end
-
-    def lstat!
-      lstat()
-    rescue SystemCallError
-      nil
-    end
-
-    def chmod(mode)
-      if symlink?
-        File.lchmod mode, path() if have_lchmod?
-      else
-        File.chmod mode, path()
-      end
-    end
-
-    def chown(uid, gid)
-      if symlink?
-        File.lchown uid, gid, path() if have_lchown?
-      else
-        File.chown uid, gid, path()
-      end
-    end
-
-    def link(dest)
-      case
-      when directory?
-        if !File.exist?(dest) and descendant_directory?(dest, path)
-          raise ArgumentError, "cannot link directory %s to itself %s" % [path, dest]
-        end
-        begin
-          Dir.mkdir dest
-        rescue
-          raise unless File.directory?(dest)
-        end
-      else
-        File.link path(), dest
-      end
-    end
-
-    def copy(dest)
-      lstat
-      case
-      when file?
-        copy_file dest
-      when directory?
-        if !File.exist?(dest) and descendant_directory?(dest, path)
-          raise ArgumentError, "cannot copy directory %s to itself %s" % [path, dest]
-        end
-        begin
-          Dir.mkdir dest
-        rescue
-          raise unless File.directory?(dest)
-        end
-      when symlink?
-        File.symlink File.readlink(path()), dest
-      when chardev?, blockdev?
-        raise "cannot handle device file"
-      when socket?
-        begin
-          require 'socket'
-        rescue LoadError
-          raise "cannot handle socket"
-        else
-          raise "cannot handle socket" unless defined?(UNIXServer)
-        end
-        UNIXServer.new(dest).close
-        File.chmod lstat().mode, dest
-      when pipe?
-        raise "cannot handle FIFO" unless File.respond_to?(:mkfifo)
-        File.mkfifo dest, lstat().mode
-      when door?
-        raise "cannot handle door: #{path()}"
-      else
-        raise "unknown file type: #{path()}"
-      end
-    end
-
-    def copy_file(dest)
-      File.open(path()) do |s|
-        File.open(dest, 'wb', s.stat.mode) do |f|
-          IO.copy_stream(s, f)
-        end
-      end
-    end
-
-    def copy_metadata(path)
-      st = lstat()
-      if !st.symlink?
-        File.utime st.atime, st.mtime, path
-      end
-      mode = st.mode
-      begin
-        if st.symlink?
-          begin
-            File.lchown st.uid, st.gid, path
-          rescue NotImplementedError
-          end
-        else
-          File.chown st.uid, st.gid, path
-        end
-      rescue Errno::EPERM, Errno::EACCES
-        # clear setuid/setgid
-        mode &= 01777
-      end
-      if st.symlink?
-        begin
-          File.lchmod mode, path
-        rescue NotImplementedError
-        end
-      else
-        File.chmod mode, path
-      end
-    end
-
-    def remove
-      if directory?
-        remove_dir1
-      else
-        remove_file
-      end
-    end
-
-    def remove_dir1
-      platform_support {
-        Dir.rmdir path().chomp(?/)
-      }
-    end
-
-    def remove_file
-      platform_support {
-        File.unlink path
-      }
-    end
-
-    def platform_support
-      return yield unless fu_windows?
-      first_time_p = true
-      begin
-        yield
-      rescue Errno::ENOENT
-        raise
-      rescue => err
-        if first_time_p
-          first_time_p = false
-          begin
-            File.chmod 0700, path()   # Windows does not have symlink
-            retry
-          rescue SystemCallError
-          end
-        end
-        raise err
-      end
-    end
-
-    def preorder_traverse
-      stack = [self]
-      while ent = stack.pop
-        yield ent
-        stack.concat ent.entries.reverse if ent.directory?
-      end
-    end
-
-    alias traverse preorder_traverse
-
-    def postorder_traverse
-      if directory?
-        entries().each do |ent|
-          ent.postorder_traverse do |e|
-            yield e
-          end
-        end
-      end
-    ensure
-      yield self
-    end
-
-    def wrap_traverse(pre, post)
-      pre.call self
-      if directory?
-        entries.each do |ent|
-          ent.wrap_traverse pre, post
-        end
-      end
-      post.call self
-    end
-
-    private
-
-    $fileutils_rb_have_lchmod = nil
-
-    def have_lchmod?
-      # This is not MT-safe, but it does not matter.
-      if $fileutils_rb_have_lchmod == nil
-        $fileutils_rb_have_lchmod = check_have_lchmod?
-      end
-      $fileutils_rb_have_lchmod
-    end
-
-    def check_have_lchmod?
-      return false unless File.respond_to?(:lchmod)
-      File.lchmod 0
-      return true
-    rescue NotImplementedError
-      return false
-    end
-
-    $fileutils_rb_have_lchown = nil
-
-    def have_lchown?
-      # This is not MT-safe, but it does not matter.
-      if $fileutils_rb_have_lchown == nil
-        $fileutils_rb_have_lchown = check_have_lchown?
-      end
-      $fileutils_rb_have_lchown
-    end
-
-    def check_have_lchown?
-      return false unless File.respond_to?(:lchown)
-      File.lchown nil, nil
-      return true
-    rescue NotImplementedError
-      return false
-    end
-
-    def join(dir, base)
-      return File.path(dir) if not base or base == '.'
-      return File.path(base) if not dir or dir == '.'
-      File.join(dir, base)
-    end
-
-    if File::ALT_SEPARATOR
-      DIRECTORY_TERM = "(?=[/#{Regexp.quote(File::ALT_SEPARATOR)}]|\\z)"
-    else
-      DIRECTORY_TERM = "(?=/|\\z)"
-    end
-
-    def descendant_directory?(descendant, ascendant)
-      if File::FNM_SYSCASE.nonzero?
-        File.expand_path(File.dirname(descendant)).casecmp(File.expand_path(ascendant)) == 0
-      else
-        File.expand_path(File.dirname(descendant)) == File.expand_path(ascendant)
-      end
-    end
-  end   # class Entry_
 
   def fu_list(arg)   #:nodoc:
     [arg].flatten.map {|path| File.path(path) }
